refactor: Tokens 일급 컬렉션 클래스로 분리
